requirements.txt
faiss-cpu==1.8.0.post1
langchain==0.3.1
langchain-community==0.3.1
langchain-core==0.3.8
langchain-text-splitters==0.3.0
numpy==1.26.4
ollama==0.3.3